refactor: create folder for regex match window
