+++ src/decorations/TextDecorationApplier.ts
@@ -1,7 +1,7 @@
 import { Range, TextDocument, TextEditor, window } from 'vscode';
 
-import { TEST_AREA_DELIMITER } from '@/FileParser';
-import RegexTest, { MatchResult } from '@/RegexTest';
+import { TEST_AREA_DELIMITER } from '@/regex-match-window/FileParser';
+import RegexTest, { MatchResult } from '@/regex-match-window/RegexTest';
 
 import { DELIMITER_DECORATION, GROUP_DECORATIONS, MATCH_DECORATION } from './utils';
