+++ src/regex-match-window/FileCreator.ts
